[qtcontacts-sqlite] Fix test failure
Signals wihtin transactions are coalesced by qtcontacts-sqlite, so
we should treat them as a set rather than as a list.

A qt4-specific error remains, where the query to match existing
aggeregates fails to find the correct candidate. It appears to be
failing due to the bound data item being corrupted in some way.