[qtcontacts-sqlite] Use lightweight aggregation relationship codepath
Previously, the "heavy" save relationships codepath was taken when
creating a new "Aggregates" relationship for a newly created contact.
In that case, the details from the new constituent had already been
promoted to the aggregate, so no regeneration is required.
In this case we gain an order of magnitude performance improvement
by simply creating the relationship via a direct query instead.

This commit also adds some useful benchmarks to the fetchtimes test.