[qtcontacts-sqlite] Improve update performance
When performing an update, constituent contacts are read as part of
the aggregation process.  This read operation is "by id" and should
be fast, however the code previously re-read all contact ids existing
in the table to ensure "existence" of the ones specified.
This commit removes that read query, and instead inserts errors
based upon the existence of the required contact in the returned
result.