[qtcontacts-sqlite] Allow syncTarget details to be modifiable
If a detail of a contact that has a syncTarget value other than 'local'
has the QContactDetail__FieldModifiable field set to true, then it
will be subject to modification during aggregate contact save.

If the aggregate has been modified such that a modifiable detail of
a syncTarget constituent has been removed, the save operation will
remove that detail from the original constituent.  Similarly, if the
aggregate has been updated to contain a modified detail and the
source detail is modifiable, the change will be applied to the source
detail.

If the relevant details are not marked as modifiable, the changes will
be applied to a local constituent.