[qtcontacts-sqlite] Always set ModifiedTimestamp on save.
Previously, the modification timestamp was only set if it was empty.
This meant that clients could explicitly tell the backend to update
the timestamp, but setting it to empty prior to save; or they could
tell the backend not to update it, by setting an explicit timestamp
manually.

However, this meant that repeated (read, update, save) operations
would result in unintuitive behaviour: the backend would not update
the timestamp automatically unless the (update) step involved the
client manually _unsetting_ the modification timestamp.

This commit changes the behaviour so that the modification timestamp
is always updated to the current datetime during save.