[qtcontacts-sqlite] Report presence changes independently
If requested via contact manager parameters, the contact manager will
report changes to contact presence details via a separate signal to
that used to report other changes.  If not requested, the usual signal
will be emitted.

A presence change is one where the only affected details (according to
the update's definition mask) are in the set { Presence, OnlineAccount,
OriginMetadata }.