[qtcontacts-sqlite] Suppress presence change notification by default
The qtcontacts-sqlite engine will now suppress the emission of the
contactsChanged signal for presence changes, unless configured to do
otherwise.  Presence changes are still reported via the internal
contactsPresenceChanged signal, which should be preferred if presence
change processing is desired.

If the 'mergePresenceChanges' parameter is set to 'true', then presence
changes will be reported via the contactsChanged signal rather than
the contactsPresenceChanged signal, which yields behavior consistent
with the Qt4 QtContacts API.  If the parameter is not supplied to the
engine, a warning is emitted to describe the effect.

The 'separatePresenceChanges' parameter is removed.