[qtcontacts-sqlite] Add function to access contact manager engine
The contact manager engine must be used to access functions which
extend the Contacts API; to access these functions, use the new
contactManagerEngine() function to retrieve a typed pointer to the
engine from the QContactManager.

Note that 'QT += contacts-private' is necessary to access the engine
property of the contact manager class.