[qtcontacts-sqlite] Provide export mode for synchronization export
Some of our data is to be restricted from general access; in order to
achieve this we will create an external copy of the database, and
export appropriately sanitized data to it. Applications that are not
authorized to access the sensitive data will use the export database
via the qtcontacts-sqlite manager.

Synchronization between the two databases will be performed using
a variant of the two-way sync adaptor interface.  The difference
between this variant and the standard mode is that the export database
will not be considered the owner of data that originates within it;
that data will become owned by the primary database once it has been
imported.  Data acquired from the export database will become property
of the local constituent in the primary database, and will not be
distinguishable from data entered directly to the primary database.