[qtcontacts-sqlite] Don't remove modified sync contacts from exportedIds
Previously, any contact which was modified remotely was removed from
the exportedIds list.  This commit changes it so that the contact
is still tracked via the exportedIds list, so that if the backend
decides that the changes are incidental, and therefore the id of the
sync contact doesn't change, we continue to receive updates for it.