[qtcontacts-sqlite] Relax permissions on created semaphores
Previously, only the uid of the process which first created the
semaphore could commit changes to the database.  This commit ensures
that any uid can do so, by relaxing the mode specified when creating
the semaphore.

Note that access to the database itself is still restricted according
to the filesystem permissions.