[qtcontacts-sqlite] Attempt to fallback to previous SHM generation
When attempting to connect to the current generation of the shared
memory region, if the current region does not exist, look for the
preceding generation.  If that exists, it is reasonably likely to be
similar to the current data, and preferable to restarting the
accumulation.

If the immediate predecessor is not available, each preceding generation
has a geomterically smaller proportion of the data, which is more
out of date - ignore the existence of these regions.