[qtcontacts-sqlite] Fix extended upgrade sequence fault
If we create a table in an upgrade statement, and subsequently alter
the same table in a later upgrade statement, the definition of the
table created in the earlier upgrade must be altered to match the
original version. Otherwise, an extended upgrade process will create
the final version of the table and then subsequently fail when the
alteration is attempted.