[qtcontacts-sqlite] Use cross join to force query order
SQLite allows manual optimization by using the 'cross join' operator,
which in addition to specifying a join instructs SQLite that it must
process the joined table after the table it is joined to.

Use the cross join operator to ensure we traverse the temporary table
of IDs before joining other data tables by index.