[qtcontacts-sqlite] Remove queryContacts internal batching
The batching inside the query processing code may provide a small
benefit or cost to overall performance; it varies from test to test
and can't be isolated.  Therefore, it is best to remove it for the
purpose of simplifying the code and requiring less maintenance of
state while processing the database records.