[qtcontacts-sqlite] Make the asynchronous DB connection primary
The contacts engine creates a synchronous DB connection at startup,
and creates an asynchronous DB connection on demand.  Clients using
libcontacts, however, will exclusively use the asynchronous
connection, so the primary connection is entirely unnecessary.

Change the startup procedure so that the asynchronous connection
becomes the primary connection and the synchronous is created as
required.