[qtcontacts-sqlite-extensions] Fix bug in delta detection in TWCSA
This commit fixes a bug in the detail delta detection of the
TwoWayContactSyncAdapter code.  Previously, the code would drop
any detail removal which was not matched with a detail addition,
while attempting to detect pairs of removals+additions as possible
modifications.  This would result in removals being "lost".
This commit ensures that all detail removals which are not matched
with a detail addition are still included in the improved delta.