[qtcontacts-sqlite] Ensure that sync resolution maintains modifiability flag
Previously, the sync resolution function could sometimes clobber the
value of the Modifiable flag on details which are modified.
This commit ensures that the value is explicitly set during every
sync resolution save operation, either to the value defined by the
sync adapter or to true if the sync adapter did not define the value.