[qtcontacts-sqlite] Improve resilience against invalid sync updates
This commit adds checks within the syncUpdate() function of the
qtcontacts-sqlite backend to improve resilience against invalid
data being provided to it by sync adapters.  Further debugging
is added to that function to allow determination of how that
invalid data is being presented.

It also adds verbose debugging to the enforceDetailConstraints
function so that when an invalid detail is encountered, that detail
and the entire contact from which it came will be dumped to the
console when QTCONTACTS_SQLITE_TRACE=1 is set.

Finally, this function ensures that detail update pairs are
generated correctly.  Previously, if it was a detail addition,
the first detail in the pair might be null (an invalid detail,
whose type is 0 = TypeUndefined).  Attempting to store that detail
would fail due to its invalid type.  This commit ensures that the
detail generated in each pair is valid (although perhaps empty).