[qtcontacts-sqlite-extensions] Allow ignorable fields to be specified in TWCSA. Contributes to MER#941
Previously, the fields which were ignored during delta detection were
hardcoded (to just the provenance and modifiability flag).

This commit allows individual sync adapters to specify which fields
of which details should be ignored when performing sync, to accommodate
sync services which don't support all possible fields.

It also improves debug logging significantly so that individual detail
differences can be displayed when QTCONTACTS_SQLITE_TWCSA_TRACE=1 is
set in the environment at runtime.

Contributes to MER#941