[qtcontacts-sqlite] Improve remote removal sync update handling. Contributes to MER#953
When multiple contacts are merged server-side (due to duplicate
detection), multiple constituents of the same aggregate may need
to be removed within the same sync operation.

This commit ensures that we deal with all server-side removal
entries in the update list, even if previous updates may have
caused modification to the (shared) aggregate of those
server-side removals.

Contributes to MER#953