[qtcontacts-sqlite] Improve handling of multiple sync target constituents. Contributes to MER#1053
Previously, the syncUpdate() function built a mapping from the
modified constituent to the appropriate sync target constituent,
in order to affect the appropriate sync target constituent instead
of the modified constituent.

The function previously did not consider the possibility that
multiple sync target constituents might be aggregated by a single
aggregate contact, and thus the mapping was incomplete.

This commit improves the mapping so that a modification to a
sync target constituent will not be re-directed to modify a
different sync target constituent (in the case where multiple
STCs exist for that contact).

It also improves the unit test to ensure that the state of the
server-side contact cache managed by the test sync adapter is
always consistent.

Contributes to MER#1053