[qtcontacts-sqlite] Allow enum property extensions to be stored. Contributes to MER#1368
Due to the schema originating to support the QtMobility.Contacts API
we store some properties in string form, when the qtpim API now uses
user-extensible enum values. Passing all values through a translation
step that only knows the predefined values causes extension values to
be lost.

To fix this problem, perform a schema upgrade in which all pre-existing
text values are converted to the equivalent enum value. In some cases
the column containing these values should be typed as INTEGER, when
in fact it is typed as TEXT. To avoid changing the columns, the read
and write routines for these columns manually translate between text
and integral form.